Change double money to integer, some fixes

+++ src/main/java/xyz/minerune/economy/provider/Provider.java
@@ -4,15 +4,15 @@ import java.util.LinkedHashMap;
 
 public interface Provider {
 
-    void create(String playerName, Double defaultMoney);
+    void create(String playerName, int defaultMoney);
 
     boolean has(String playerName);
 
-    Double get(String playerName);
+    int get(String playerName);
 
-    void set(String playerName, Double amount);
+    void set(String playerName, int amount);
 
-    LinkedHashMap<String, Double> getAll();
+    LinkedHashMap<String, Integer> getAll();
 
     String getName();
 }
